drivers: dma: pl330: add peripheral support
dma-cells now should have the peripheral id in the dts.
In the case of mem2mem, it can be set to 0.

The opcode generation logic is changed to incorporate
the fixed burst size requirement of peripherals.

Assign the req_type based on the burst_len requested.

Validate burst length and return appropriate error.

Signed-off-by: Sudhir Sreedharan <sudhir@alifsemi.com>
Signed-off-by: Tanay Rami <tanay@alifsemi.com>