soc: arm: alif_ensemble: linker: include alif sections
There are instances where the symbols should be placed
in alif sram sections before being placed in the default
noinit sections.

For example, by default, the video buffer is placed in the
noinit section of dtcm, but for alif, this buffer needs to be placed
in the sram noinit section.

Signed-off-by: Sudhir Sreedharan <sudhir@alifsemi.com>