boards: alif_e7_dk: Add support for ae722f80f55d5xx_apss board
Add support for alif ensemble e7 devkit with support for the
cortex-a32 application processor.

The current ensemble SOC device tree files are tightly coupled
with armv8-m components. So a standalone device tree was chosen
for now for the apss board so as not to disturb the rtss boards
and other device tree include files. The board currently supports
the GIC, arm arch timer, alif clock control driver, alif pinctrl
driver and the UARTs.

Signed-off-by: Silesh C V <silesh@alifsemi.com>