drivers: mspi_dw: Edit core disable to after PM init
There is a scenario where a Bus fault occurs as the
power management isn't initialised yet and the core
is attempted to be turned off via the ssienr register.
This commit edits this so the core register is written
to after the initialisation of the power management.

Signed-off-by: David Jewsbury <david.jewsbury@nordicsemi.no>
(cherry picked from commit 4bf6a756280be4d90edf96f1a4c088c4620be41d)
Signed-off-by: Ganesh Kumar R <ganesh.ramani@alifsemi.com>