soc: arm: ensemble: linker: Include alif sections
Based on logic introduced in commit
2fa51a201fef7a59c71e5132f7d3ba2f5bc76c9e, ZAS 1.5 branch.

There are instances where the symbols should be placed
in alif sram sections before being placed in the default
noinit sections.

For example, by default, the video buffer is placed in the
noinit section of dtcm, but for alif, this buffer needs to be placed
in the sram noinit section.

Signed-off-by: Yogender Kumar Arya <yogender.kumar@alifsemi.com>